Move the cli lint command behind a main option
